unity_backend/requirements.txt
coverage==5.4
crypto==1.4.1
cryptography==3.4.7
Django==3.1.8
django-appconf==1.0.4
django-autoslug==1.9.8
django-compressor==2.4
django-debug-toolbar==3.2.1
django-extensions==3.1.3
django-filter==2.4.0
django-haystack==3.0
django-ipware==3.0.2
django-mdeditor==0.1.18
django-uuslug==1.2.0
django-werkzeug-debugger-runserver==0.3.1
djangorestframework==3.12.4
Naked==0.1.31
netifaces==0.10.4
oauthlib==3.1.0
pycrypto==2.6.1
pycryptodome==3.10.1
PyMySQL==1.0.2
PyNaCl==1.3.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyrsistent==0.17.3
pytz==2021.1
raven==6.10.0
rcssmin==1.0.6
requests==2.25.1
rjsmin==1.1.0
setuptools-scm==6.0.1
shellescape==3.8.1
six==1.15.0
sqlparse==0.4.1
ssh-import-id==5.10
text-unidecode==1.3
toml==0.10.2
